Review feedback; rename package, types, and methods
